+++ lapdog/__init__.py
@@ -1,3 +1,3 @@
 from dalmatian import *
 from .lapdog import *
-__version__ = "0.13.4"
+__version__ = "0.13.5"