+++ lapdog/operations.py
@@ -100,7 +100,7 @@ class Operator(object):
             traceback.print_exc()
         self.live = True # To halt background thread
 
-    def synchronized(func):
+    def __synchronized(func):
         def wrapper(self, *args, **kwargs):
             with self.lock:
                 return func(self, *args, **kwargs)
@@ -145,7 +145,7 @@ class Operator(object):
         with self.lock:
             self._thread = None # Delete it's own reference
 
-    @synchronized
+    @__synchronized
     def go_live(self):
         failures = []
         exceptions = []
@@ -180,7 +180,7 @@ class Operator(object):
         self.live = not len(self.pending)
         return self.live, exceptions
 
-    @synchronized
+    @__synchronized
     def tentative_json(self, result, *expected_failures):
         self.last_result = result
         if result.status_code >= 400 and result.status_code not in expected_failures:
@@ -233,7 +233,7 @@ class Operator(object):
         self.fail()
 
     @property
-    @synchronized
+    @__synchronized
     def firecloud_workspace(self):
         if self.live:
             with self.timeout('workspace'):
@@ -248,7 +248,7 @@ class Operator(object):
         self.fail()
 
     @property
-    @synchronized
+    @__synchronized
     def configs(self):
         if self.live:
             with self.timeout('configs'):
@@ -264,7 +264,7 @@ class Operator(object):
             return self.cache['configs']
         self.fail()
 
-    @synchronized
+    @__synchronized
     def get_config_detail(self, namespace, name):
         key = 'config:%s/%s' % (namespace, name)
         if self.live:
@@ -281,7 +281,7 @@ class Operator(object):
             return self.cache[key]
         self.fail()
 
-    @synchronized
+    @__synchronized
     def _upload_config(self, config):
         configs = self.configs
         if (config['namespace']+'/'+config['name']) not in ['%s/%s'%(m['namespace'], m['name']) for m in configs]:
@@ -302,7 +302,7 @@ class Operator(object):
                 print(r.text)
         return False
 
-    @synchronized
+    @__synchronized
     def add_config(self, config):
         key = 'config:%s/%s' % (config['namespace'], config['name'])
         if 'configs' in self.cache and self.cache['configs'] is not None:
@@ -340,7 +340,7 @@ class Operator(object):
         ))
         return False
 
-    @synchronized
+    @__synchronized
     def get_wdl(self, namespace, name, version=None):
         key = 'wdl:%s/%s' % (namespace, name)
         if version is not None:
@@ -364,7 +364,7 @@ class Operator(object):
             return self.cache[key]
         self.fail()
 
-    @synchronized
+    @__synchronized
     def upload_wdl(self, namespace, name, synopsis, path, delete=True):
         key = 'wdl:%s/%s' % (namespace, name)
         with open(path) as r:
@@ -393,7 +393,7 @@ class Operator(object):
                 partial(self.call_with_timeout, DEFAULT_LONG_TIMEOUT, dog.get_wdl, namespace, name)
             ))
 
-    @synchronized
+    @__synchronized
     def validate_config(self, namespace, name):
         cfgkey = 'config:%s/%s' % (namespace, name)
         key = 'valid:'+cfgkey
@@ -421,7 +421,7 @@ class Operator(object):
         }
 
     @property
-    @synchronized
+    @__synchronized
     def entity_types(self):
         if self.live:
             with self.timeout('entity_types'):
@@ -438,7 +438,7 @@ class Operator(object):
         self.fail()
 
     @property
-    @synchronized
+    @__synchronized
     def _entities_live_update(self):
         state = self.live
         self.live = True
@@ -447,7 +447,7 @@ class Operator(object):
         finally:
             self.live = state
 
-    @synchronized
+    @__synchronized
     def get_entities_df(self, etype):
         getter = partial(
             getattr(
@@ -470,7 +470,7 @@ class Operator(object):
             return self.cache[key]
         self.fail()
 
-    @synchronized
+    @__synchronized
     def update_entities_df(self, etype, updates, index=True):
         getter = partial(
             self.call_with_timeout,
@@ -542,7 +542,7 @@ class Operator(object):
             except APIException:
                 pass
 
-    @synchronized
+    @__synchronized
     def update_entities_df_attributes(self, etype, updates):
         getter = partial(
             self.call_with_timeout,
@@ -613,7 +613,7 @@ class Operator(object):
             except APIException:
                 pass
 
-    @synchronized
+    @__synchronized
     def update_entity_set(self, etype, set_id, member_ids):
         key = 'entities:%s_set' % etype
         updates = pd.DataFrame(index=[set_id], data={etype+'s':[[*member_ids]]})
@@ -688,7 +688,7 @@ class Operator(object):
             return ws['workspace']['attributes']
         self.fail()
 
-    @synchronized
+    @__synchronized
     def update_attributes(self, attrs):
         if 'workspace' in self.cache:
             if self.cache['workspace'] is None:
@@ -722,7 +722,7 @@ class Operator(object):
             except AssertionError:
                 pass
 
-    @synchronized
+    @__synchronized
     def evaluate_expression(self, etype, entity, expression):
         if self.live:
             with self.timeout(DEFAULT_LONG_TIMEOUT):